# [logger] Fix line number wrapping bug for files with >999 lines