--- a/esphome/components/logger/logger.h
+++ b/esphome/components/logger/logger.h
@@ -355,12 +355,18 @@ class Logger : public Component {
     buffer[pos++] = '[';
     copy_string(buffer, pos, tag);
     buffer[pos++] = ':';
-    int hundreds = line / 100;
-    int remainder = line - hundreds * 100;
-    int tens = remainder / 10;
-    buffer[pos++] = '0' + hundreds;
-    buffer[pos++] = '0' + tens;
-    buffer[pos++] = '0' + (remainder - tens * 10);
+    if (line > 999) {
+      buffer[pos++] = 'B';
+      buffer[pos++] = 'I';
+      buffer[pos++] = 'G';
+    } else {
+      int hundreds = line / 100;
+      int remainder = line - hundreds * 100;
+      int tens = remainder / 10;
+      buffer[pos++] = '0' + hundreds;
+      buffer[pos++] = '0' + tens;
+      buffer[pos++] = '0' + (remainder - tens * 10);
+    }
     buffer[pos++] = ']';
 
 #if defined(USE_ESP32) || defined(USE_LIBRETINY) || defined(USE_ZEPHYR)