# What does this implement/fix?

Fixes incorrect line number formatting for files with >999 lines. Line numbers were being corrupted (displaying garbage characters) or incorrectly wrapped. This PR properly formats line numbers up to 9999 without using modulo operations.

## Problem

#10960 introduced a regression for line numbers > 999:

**Before #10960 (original snprintf version):**
```cpp
this->printf_to_buffer_(buffer, buffer_at, buffer_size, "%s[%s][%s:%03u]: ", color, letter, tag, line);
```
- Used `%03u` format which correctly handled all line numbers
- Line 1389 displayed correctly as "1389" (4 digits)
- No bug, but slower performance

**After #10960 (current modulo-based optimization):**
```cpp
buffer[pos++] = '0' + (line / 100) % 10;
buffer[pos++] = '0' + (line / 10) % 10;
buffer[pos++] = '0' + line % 10;
```
- 35-72% performance improvement by avoiding snprintf
- Introduced bug: Line numbers > 999 wrap around using modulo
- Line 1389 displays as "389" - misleading for debugging (looks like line 389)

## Solution

This PR properly handles line numbers up to 9999 by adding a 4th digit when needed, using subtraction instead of modulo to avoid the overhead:

```cpp
// Format line number without modulo operations (passed by value, safe to mutate)
if (line > 999) [[unlikely]] {
  int thousands = line / 1000;
  buffer[pos++] = '0' + thousands;
  line -= thousands * 1000;
}
int hundreds = line / 100;
int remainder = line - hundreds * 100;
int tens = remainder / 10;
buffer[pos++] = '0' + hundreds;
buffer[pos++] = '0' + tens;
buffer[pos++] = '0' + (remainder - tens * 10);
```

**Output examples:**
- Line 89 → `[tag:089]` (3 digits, unchanged)
- Line 999 → `[tag:999]` (3 digits, unchanged)
- Line 1000 → `[tag:1000]` (4 digits, now correct!)
- Line 1389 → `[tag:1389]` (4 digits, now correct!)

## Flash Impact

**ESP8266:**
- Adds 32 bytes flash (351035 → 351067 bytes)
- Trade-off: 32 bytes for correct line number display on large files

The extra flash is needed for the 4-digit formatting logic, but this is necessary to fix the corruption bug.

## Performance

As a bonus, the subtraction-based approach (even with 4-digit support) avoids expensive modulo operations:

Benchmarked on host with `-Os` optimization (matching ESPHome build settings, 1M iterations per run, 50 runs):

| Test Case | Original (snprintf) | Modulo-based (#10960) | This PR (no modulo + 4-digit) | Improvement vs modulo |
|-----------|---------------------|----------------------|-------------------------------|----------------------|
| INFO level, short tag | 82.6 ns ± 3.8 ns | 6.6 ns ± 0.0 ns | 6.3 ns ± 0.1 ns | **3.6% faster** |
| ERROR level, long tag | 77.3 ns ± 2.4 ns | 6.3 ns ± 0.1 ns | 6.1 ns ± 0.2 ns | **4.1% faster** |
| DEBUG level | 76.9 ns ± 2.5 ns | 5.8 ns ± 0.0 ns | 5.5 ns ± 0.0 ns | **5.1% faster** |
| VERY_VERBOSE level | 79.6 ns ± 2.5 ns | 6.0 ns ± 0.0 ns | 6.0 ns ± 0.0 ns | **0.3% faster** |
| With thread name (ESP32) | 76.8 ns ± 1.7 ns | 13.0 ns ± 0.3 ns | 12.7 ns ± 0.2 ns | **2.0% faster** |

## Testing

**Correctness verification:**
- Line numbers 0-999: Identical output to previous version
- Line numbers 1000-9999: Now display correctly with 4 digits
- Tested edge cases: 89, 999, 1000, 1389, 9999

**Real-world example:**
- `api_connection.cpp:1389` now displays as `[api.connection:1389]` instead of `[api.connection:=89]`

## Types of changes

- [x] Bugfix (non-breaking change which fixes an issue)
- [x] Code quality improvements to existing code or addition of tests

**Related issue or feature (if applicable):**

- Fixes line number display bug introduced/exposed by #10960
- Files with >999 lines (like `api_connection.cpp`) were showing corrupted line numbers

**Pull request in [esphome-docs](https://github.com/esphome/esphome-docs) with documentation (if applicable):**

- N/A - internal bug fix, no user-facing configuration changes

## Test Environment

- [x] ESP32
- [x] ESP32 IDF
- [x] ESP8266
- [ ] RP2040
- [ ] BK72xx
- [ ] RTL87xx
- [ ] nRF52840

## Example entry for `config.yaml`:

```yaml
# No configuration changes - this is an internal bug fix
# All existing logger configurations work identically, now with correct line numbers
```

## Checklist:
  - [x] The code change is tested and works locally.
  - [x] Tests have been added to verify that the new code works (under `tests/` folder).

If user exposed functionality or configuration variables are added/changed:
  - [ ] Documentation added/updated in [esphome-docs](https://github.com/esphome/esphome-docs).