@bdraco 2025-10-02
So for lines > 999 it fails. modulo failed differently. We should probably handle this case