@bdraco 2025-10-02
all good now [00:47:27.987][D][api.connection:1389]: aioesphomeapi (192.168.113.95) connected